spec: rename files back to spec.c and spec.h
